Add grid instance variable to the grid artist
This is purely a refactor commit, the behavior remains unchanged.

This commit changes the way the Grid2DArtist works. It used to accept a
grid each time the draw() function was called. Now, the grid is
specified on initialization and the draw() command just uses the
instance variable.

This change is made because we need to add click-handling to the artist,
and for that we need to have a reference to the current grid.