Add interactive grid
This commit adds the ability to add and remove walls by clicking on
cells in the grid. This dynamically updates the path finding and
redraws.

To make this happen I've added the ability for the artist to handle
click events.

Via a callback, the demo.js can now redraw and find a new path each time
a cell is clicked.